Splitting the forwarder to have one per domain.
The forwarder class has been split in two: forwarder and
domainForwarder.

We now spawn one "domainForwarder" per domain in the configuration. This
avoids slowing down one domain when the another one is down.

We also split the health report for the forwarder to its own class to
ease testing.

Test coverage has been pushed from 78.4% to 92.4% and the README has been
updated.